requirements.txt
argon2-cffi==21.3.0
asyncpg==0.27.0
cryptography==38.0.4
decorator==5.1.1
fastapi==0.88.0
fastapi-pagination==0.11.1
python-jose==3.3.0
python-multipart==0.0.5
tortoise-orm==0.19.2
uvicorn==0.20.0
